pyproject.toml
[project]
name = "clinical-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "qdrant-client[fastembed]>=1.11.0",
    "httpx>=0.27.0",
    "pandas>=2.2.2",
    "pyarrow>=17.0.0",
    "numpy>=1.26.4",
    "pydantic>=2.7.0",
    "openai>=1.40.0",
    "streamlit>=1.36.0",
    "scikit-learn>=1.4.2",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.1",
    "pytest>=8.4.2",
    "nbformat>=5.10.4",
    "ipykernel>=6.30.1",
    "nbclient>=0.10.2",
    "great-tables>=0.18.0",
]
